Extract all actions to their own submodule. Refactor imports on main file
