+++ ti/actions/write/note.py
@@ -1,0 +1,19 @@
+from ti.dataaccess.utils import get_data_store
+from ti.actions.utils.utils import ensure_working
+
+
+def action_note(colorizer, content):
+    data = get_data_store('JSON').load()
+
+    ensure_working(data)
+
+    current = data['work'][-1]
+
+    if 'notes' not in current:
+        current['notes'] = [content]
+    else:
+        current['notes'].append(content)
+
+    get_data_store('JSON').dump(data)
+
+    print('Yep, noted to ' + colorizer.yellow(current['name']) + '.')
